Import py-fido2
Provides library functionality for communicating with a FIDO device
over USB as well as verifying attestation and assertion signatures.
This library aims to support the FIDO U2F and FIDO 2.0 protocols for
communicating with a USB authenticator via the Client-to-Authenticator
Protocol (CTAP 1 and 2). In addition to this low-level device access,
classes defined in the fido2.client and fido2.server modules implement
higher level operations which are useful when interfacing with an
Authenticator, or when implementing WebAuthn support for a Relying
Party.

As warned by upstream, "This project is in beta. Expect things to
change or break at any time!" - it is currently known not to work on
OpenBSD with some device types.


From Lucas Raab < tuftedocelot at fastmail dot fm >, thanks!
Feedback from Georg Steuck
Feedback and OK kmos (earlier version)
Feedback and OK sthen