import sysutils/snmp_exporter, ok tb@ espie@
Prometheus exporter for metrics gathered from SNMP. Contains an exporter
that does the actual scraping, and a generator (which depends on Net-SNMP)
that creates the configuration for use by the exporter.

The snmp exporter reads from a snmp.yml config file by default. This file
is not intended to be written by hand, rather use the generator to generate
it for you.