Remove KDE4 baseapps
Our KDE4 baseapps port contains only konqueror. It's no longer usable and the
only good thing about konqueror4 is that it doesn't crash.

It is not necessary to display the KDE4 help pages so it's time to say good bye!
Dear KDE4 browser users, there are enough alternatives to visit the web or your
file system in OpenBSD.