Update to 6.22.02. 6.21 and 6.22 are bug fix releases with no new features.
Stop installing the message catalogs, which were already broken.
tcsh uses the obscure POSIX catopen() interface, which requires the
environment variable NLSPATH to be set to access the translations.
NLSPATH is not set by default on OpenBSD and users are unlikely to
figure out the necessary magic.  Discussed with schwarze@