Import envconsul-0.9.1.
Envconsul provides a convenient way to launch a subprocess with environment
variables populated from HashiCorp Consul and Vault. The tool is inspired by
envdir and envchain, but works on many major operating systems with no runtime
requirements. It is also available via a Docker container for scheduled
environments
<...>

ok robert@