Import stixfonts-2.0.2.
The Scientific and Technical Information eXchange (STIX) fonts are intended
to meet the demanding needs of authors, publishers, printers, and others in
the scientific, medical, and technical fields. They combine a comprehensive
Unicode-based collection of mathematical symbols and alphabets with a set of
text faces suitable for professional publishing.

From George Rosamond; thanks!

ok kmos@