devel/py-test-timeout
"This is a plugin which will terminate tests after a certain timeout.
When doing so it will show a stack dump of all threads running at
the time. This is useful when running tests under a continuous
integration server or simply if you don't know why the test suite
hangs."

Needed to enable tests on (at least) www/py-aiohttp

Version 1.3.3

OK kn@