asterisk port tweaks:
- use getthrid to record TID (this is displayed following the opaque thread
id in "core show threads", and is useful to match against top -H output)

- provide a -kqueue package for res_timing_kqueue (previously @comment'ed
out). Normally a pthread-backed timer is used in the package; the kqueue-backed
timer is built by default but has had problems on OpenBSD in the past so is
@comment'ed out. I wouldn't consider this production ready on OpenBSD (lots
of "kqueue_timer_ack: [18]: Missed 1" at least on a kernel with standard HZ)
but I'd like to have it more easily available for experimentation, hence
adding the package.