import open62541 1.0.1 OK sthen@
Comment:
library implementation of OPC UA

Description:
open62541 is an open source and free implementation of OPC UA (OPC
Unified Architecture) written in the common subset of the C99 and
C++98 languages.  The library is usable with all major compilers
and provides the necessary tools to implement dedicated OPC UA
clients and servers, or to integrate OPC UA-based communication
into existing applications.