update py-netmiko to 3.0.0
this drops support for python2, so adjust dependant ports accordingly (py-IOSXR
and py-napalm). as discussed with stuart we will force the python3 flavor for
these ports to make it easier in the future to add additional flavors when needed.

add quirks as needed and remove the non-python3 entries from the category Makefiles

ok sthen@