Import devel/esptool-ck: firmware creation helper for esp8266 chips
Esptool is a tool to create firmware files for the ESP8266 chip and flash the
firmware to the chip over serial port. Esptool reads the compiled program in ELF
format, extracts code and data sections, and either dumps a section to a file or
assembles the firmware file from several segments. Esptool also communicates
with the ESP8266 bootloader to upload firmware files to flash.

With some corrections from martin@ and kmos@

ok martin@ and kmos@