xz: change from sysctl hw.ncpu to sysconf(_SC_NPROCESSORS_ONLN), ok naddy@
This is used for -T0 which auto-detects the number of cores for parallel
compression. xz prefers sysctl over sysconf (see m4/tuklib_cpucores.m4
and src/common/tuklib_cpucores.c for reasons) but this doesn't work for
us; just set an autoconf cache variable to force sysconf which works
better for us.