Import net/dbip.
The free DB-IP IP to Country and IP to City Lite databases are a subset of
the IP to Location databases with reduced coverage and accuracy distributed
under the Creative Commons Attribution 4.0 International License (CC BY 4.0).

These databases are provided in MMDB format which can be directly consumed
with net/libmaxminddb.

OK jca@