Import py-sip-qt5 4.19.19.
SIP is a tool for automatically generating Python bindings for C and C++
libraries. This module contains Qt5 bindings that were previously in py-qt5.

not yet linked to the build but will be necessary for the upcoming
py-qt5 update. Not importing 12.7.x versions as they are python3 only,
and we still have some consumers for the python2 FLAVOR of py-qt5.

to be tightly kept in sync with devel/py-sip, and tested against py-qt5
updates..

with tweaks from and ok sthen@