import ports/x11/tigervnc, ok robert@ bcallah@
TigerVNC is a high-performance, platform-neutral implementation of VNC
(Virtual Network Computing), a client/server application that allows
users to launch and interact with graphical applications on remote
machines.

TigerVNC provides the levels of performance necessary to run 3D and
video applications, and it attempts to maintain a common look and feel
and re-use components, where possible, across the various platforms
that it supports. TigerVNC also provides extensions for advanced
authentication methods and TLS encryption.

This package provides TigerVNC's viewer and "x0vncserver", which shares
an existing X server (typically, one that is connected to a physical
screen) with viewers on the network.