import ports/devel/xtensa-esp32-elf, ok sthen@ kmos@
The xtensa-esp32-elf port is an ESP32 GNU cross-compiler Toolchain.

ESP32 is a series of low-cost, low-power system on a chip microcontrollers with
integrated Wi-Fi and dual-mode Bluetooth. The ESP32 series employs a Tensilica
Xtensa LX6 microprocessor in both dual-core and single-core variations and
includes built-in antenna switches, RF balun, power amplifier, low-noise receive
amplifier, filters, and power-management modules.

ESP32 is created and developed by Espressif Systems, a Shanghai-based Chinese
company, and is manufactured by TSMC using their 40 nm process.[2] It is a
successor to the ESP8266 microcontroller.