Register libcblas.so dependency on libblas.so
This avoids a runtime error with software linked by ld.lld -lcblas, and
a build error with ld.bfd -lcblas.

Spotted by sthen@ and tb@ while working on a py-numpy update.  Tests by
tb@ and martin@, ok tb@

A bunch of people want to move forward with this, so I'm not waiting for
Steven's ok.