Remove tortoisehg.
Tortoisehg uses the internal mercurial API. We must to maintain mercurial and
tortoisehg in sync. In the last year, the releases have been lagging, sometimes
for months, only published as a tag in bitbucket, not always and not at time
for mayor mercurial versions. On the other hand, mercurial is a server exposed
to the network using an unsupported python version. Our actual mercurial
version would be 1+ year old for 6.8 with large changes in between (python2 to
python3).

We could add the port again in the future if upstream starts to release new
versions regularly.

OK paco@ gkoehler@. No objection sthen@.