Hack through upstream's Makefile to build libtinyxml.so
Instead of a handrolled do-build that doesn't respect COMPILER,
resulting in a library built with base-g++ on sparc64 and friends.
That doesn't fly with other ports being built with C++11 and ports-gcc.
shared lib version major bump because of the resulting export changes on
base-gcc archs, no change with base-clang.

This should fix at least
http://build-failures.rhaalovely.net/sparc64/2020-03-08/games/pokerth.log

This also adds a dedicated test target and drops NO_TEST.