Disable use libexecinfo unconditionally
tmate has no LIB_DEPENDS on devel/libexecinfo and the configure script does
not detect it even if installed.

Since all of tmate-debug.c is under an HAVE_BACKTRACE macro guard, it makes
no sense to compile most of it and only keep calls to libexecinfo under a
DEBUG macro guard.

Disable it entirely to simplify the patch and avoid dead code.