Remove audio/nap.
nap has been dead upstream for years, and the latest version is 1.5.4
from 2006. The version we package is 1.5.3 from 2004.

Debian deleted it [1] in 2017, and besides us, only Pkgsrc still have a
package for it.

The server list it is trying to fetch is unavailable and the few servers
listed on http://nap.sourceforge.net are not available anymore.

OK bcallah@, kmos@, jca@

[1] https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=867764