Drop the post-install "if you want to use this as your default system Python" messages, Python 2 development has finished so this is not a sensible option to use as default.
(It is still kept in the ports tree for now, as some important software
has not been updated to use Python 3).

ok tracey aja mariani rpointel

From https://www.python.org/doc/sunset-python-2/, "We are volunteers who
make and take care of the Python programming language. We have decided
that January 1, 2020, was the day that we sunset Python 2. That means
that we will not improve it anymore after that day, even if someone
finds a security problem in it. You should upgrade to Python 3 as soon
as you can."