Delete mail/akpop3d.
Alexei dot Malinin at mail dot ru reported a compiler warning that,
in my opinion, probably indicates a security vulnerability, but due
to an incomplete description of the affected feature in the
documentation, it is unclear how it should be fixed.  The program
appears to be sloppily written, sloppily documented, and abandoned
upstream 15 years ago.

OK ajacoutot@ for deleting it.