pdflib: fix the build on powerpc
In various places in the code, it's assumed that the __POWERPC__ define
is Mac OS 9 specific, but our clang defines it as well, breaking the build,
revert to the gcc situation.

Also add a license marker (thanks to jca@ for figuring out under which
license we should distribute it), and remove MASTER_SITES because it
can't be fetched from upstream; this port may be removed soon or
otherwise its distfile mirrored.

OK sthen@ and jca@