Import p5-Test-Spec-0.54
This is a declarative specification-style testing system for
behavior-driven development (BDD) in Perl. The tests (a.k.a. examples)
are named with strings instead of subroutine names, so your fingers
will suffer less fatigue from underscore-itis, with the side benefit
that the test reports are more legible.

This module is inspired by and borrows heavily from RSpec, a BDD
tool for the Ruby programming language.

OK sthen@