shotcut is a full-featured video editor
loosely based on WIP by bentley@

the current port is not finished, but it can do a few things: play videos,
cut tracks, export the results.

A lot of the filters will currently crash! So not linked to the build yet.

Video fade-in/fade-out do not do anything yet

Imported so that it's easier to figure out the issues in-tree.

okay tracey@ rsadowski@