Update to Subversion 1.14.0.
Biggest news in this release is support for Python3 in the bindings.
Keep installing the Python2 bindings for now as there are other ports
which still depend on them, such as devel/cvs2svn and www/trac.

With tweaks from sthen@ who also tested www/trac with SVN 1.14.

ok sthen@