Import reSIProcate stack, repro, and reTurn 1.12.0
The reSIProcate components, particularly the SIP stack, are in use in both
commercial and open-source products. The project is dedicated to maintaining a
complete, correct, and commercially usable implementation of SIP and a few
related protocols.

repro is an open-source, free SIP server which provides SIP proxy, registrar,
redirect, and identity services.

reTurn is a highly efficient C++ open-source STUN/TURN server and client
library. It is an implementation of the latest STUN/TURN RFCs: RFC5389 (STUN),
and RFC5766 (TURN).

OK sthen@