Switch to using autoconf and patching aclocal.m4 and configure.ac.
This allows for cleaning up the Makefile and for at least a chance
to get things like -z notext and -z wxneeded merged upstream for
amd64 *and* i386.

It also removes a -fno-pie hidden in patches/patch-configure.

Looks good to Greg, who also tested it by building xmonad and all
of its dependencies using cabal (in addition to my ordinary build
tests).