gnome-builder: fix the build with ld.bfd
Don't use "-Wl,--as-needed" (the meson default) with ld.bfd, it causes
undefined references errors. Allow building with ld.bfd on i386 where
it's broken with ld.lld, we'll see the result in the next bulk.

Tested on powerpc -- this arch needs a yet-to-be-committed
x11/gnome/libdazzle fix to build it there.

OK sthen@, jca@ and aja@ (maintainer)