Install akregator-19.12.3
Comment:
news feed reader

Description:
Akregator is a news feed reader. It enables you to follow news sites, blogs and
other RSS/Atom-enabled websites without the need to manually check for updates
using a web browser. Akregator is designed to be both easy to use and to be
powerful enough to read hundreds of news sources conveniently. It comes with a
fast search, advanced archiving functionality and an internal browser for easy
news reading.

Maintainer: Rafael Sadowski <rsadowski@openbsd.org>