Import devel/py-dill.
dill extends python's pickle module for serializing and de-serializing
python objects to the majority of the built-in python types.
Serialization is the process of converting an object to a byte
stream, and the inverse of which is converting a byte stream back
to on python object hierarchy.

Required by py-multiprocess