Re-import math/ebc as math/bc-gh. Renaming because of upstream preference for this name, plus convergence with other packaging systems that use this name.
While here, update bc-gh to version 3.1.1. Also improve 64-bit check in the
Makefile.

ok jca@ some time ago

This is an implementation of the POSIX bc calculator that implements GNU
bc extensions as well as the period (.) extension from the BSD flavor of
bc.

This bc also includes an implementation of dc in the same binary,
accessible via a symbolic link, which implements all *BSD and GNU
extensions, though the ! command is omitted for security concerns.