net/aria2: fix after enabling TLSv1.3 for TLS_method()
The port wants to know what version of TLS it is using and asserts that
it knows. Unsurprisingly, the mechanism depends on TLS1_3_VERSION being
defined to the correct value.  We cannot define this publicly in libssl
yet since this would result in a lot of breakage in other ports, so use
a local define in the build environment.

Reported and tested by Martin (martingot () protonmail com)

ok gonzalo (maintainer), kn