Bump libclang.so and friends
libclang.so is used by other ports and the update to llvm 10 warranted
a minor bump.  I haven't checked liblldb.so and libLTO.so, I bumped
their major version just in case but they aren't used in the ports tree.

Thanks to kn@ who provided an old libclang.so for comparison.