Add a test for zero content type from jsing
This checks whether a TLSv1.3 server handles the case of a record
with all-zero plaintext correctly (i.e., by sending an unexpected
message alert) in various contexts.  This would have caught a bug
where the LibreSSL TLSv1.3 record layer would call freezero() with
a size of -1.

What's missing is a test that sends a completely empty record. The
framework doesn't seem to permit that easily.