Import x11/hsetroot, a wallpaper changer that works with compositors. Original submission from Stefan Hagen, who takes MAINTAINER -- thanks! tweak and ok sthen@
hsetroot is a tool which allows you to compose wallpapers ("root
pixmaps") for X. It has a lot of options like rendering gradients,
solids and images and manipulate these in various ways. hsetroot also
supports alpha-channels.

hsetroot uses XSetWindowBackgroundPixmap instead of XSetWindowBackground
for solid colors. This makes it a good xsetroot replacement when a
compositor (compton, xcompmgr, ...) is in use.