Update to Redis 6.0.7
Upgrade urgency for both 6.0.6 and 6.0.7 is MODERATE, see changelog for
details. Notably, the fix for the sentinel crash was merged upstream.

https://raw.githubusercontent.com/antirez/redis/6.0/00-RELEASENOTES

There is another test that occasionally fails since 6.0.6:

[err]: Piping raw protocol in tests/integration/redis-cli.tcl

Tested on amd64 and macppc, build tested with clang and egcc on sparc64.

Thanks to danj for testing and okaying 6.0.6.

ok sthen