import shockolate, a source port of the engine for System Shock ok solene@ (minor tweaks to README since then)
DESCR:
Shockolate - System Shock, but cross platform! Based on the source code for
PowerPC released by Night Dive Studios, Incorporated. Shockolate is a cross
platform source port of System Shock, using SDL2. This runs well on OSX, Linux,
and Windows right now, with some missing features that need reviving due to not
being included in the source code that was released.

The end goal for this project is something like what Chocolate Doom is for Doom:
an experience that closely mimics the original, but portable and with some
quality of life improvements including an OpenGL renderer and mod support!